REFACTOR:used Flaot instead of Double

+++ src/main/java/com/generics/MaximumValueClass.java
@@ -11,8 +11,8 @@ public class MaximumValueClass {
         return max;
     }
 
-    public Double findMaximum(Double a, Double b, Double c) {
-        Double max=a;
+    public Float findMaximum(Float a, Float b, Float c) {
+        Float max=a;
         if(b.compareTo(max)>0)
             max=b;
         if(c.compareTo(max)>0)
